refactor: extract avatar-group logic into reusable mixin
