+++ packages/avatar-group/package.json
@@ -36,6 +36,7 @@
     "polymer"
   ],
   "dependencies": {
+    "@open-wc/dedupe-mixin": "^1.3.0",
     "@polymer/polymer": "^3.0.0",
     "@vaadin/a11y-base": "24.6.0-alpha8",
     "@vaadin/avatar": "24.6.0-alpha8",
